Safari-609.1.11.1/LayoutTests/fast/text-indicator/text-indicator-uses-img-size-inside-picture-expected.txt


imageOnly: -2 -1 260 258
picture: -2 -1 260 258
clippedPicture: -2 -1 260 22
deeperClippedPicture: -2 -1 260 22
